pyproject.toml
[project]
name = "emu"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1",
    "transformers==4.44",
    "tiktoken==0.6.0",
    "pillow",
    "gradio==4.44.0",
]
packages = [{ include = "emu", from = "src" }]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
include = ["*.py"]
dev-mode-dirs = [".", "src"]

[tool.hatch.build.sources]
"src/emu" = "emu"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["emu"]